+++ monopatin-microservicio/src/main/java/com/monopatinmicroservicio/service/DTO/ReporteKilometrajeDTO.java
@@ -5,12 +5,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReporteKilometrajeDTO implements Serializable {
-    private double kilometros_recorridos;
-    private int minutos_uso;
+    private Double kilometros_recorridos;
+    private BigDecimal minutos_uso;
     private Long cantidad_viajes;
+
 }
